Refactored upsert / delete  into separate classes
