+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/IUpsertGraphSyncer.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using OrchardCore.ContentManagement;
+
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers
+{
+    public interface IUpsertGraphSyncer
+    {
+        Task SyncToGraph(ContentItem contentItem);
+    }
+}
